reverted to just redis
