--- tests/config/state/aerospike/statestore.yaml
@@ -1,11 +1,0 @@
-apiVersion: dapr.io/v1alpha1
-kind: Component
-metadata:
-  name: aerospike
-spec:
-  type: state.Aerospike
-  metadata:
-  - name: hosts
-    value: "localhost:3000"
-  - name: namespace
-    value: conformance
