--- tests/config/state/cassandra/statestore.yaml
@@ -1,9 +1,0 @@
-apiVersion: dapr.io/v1alpha1
-kind: Component
-metadata:
-  name: cassandra
-spec:
-  type: state.cassandra
-  metadata:
-  - name: hosts
-    value: localhost:9042
