--- tests/results/state/aerospike.json
@@ -1,43 +1,0 @@
-{
-    "conformant": true,
-    "component_type": "state",
-    "component_name": "aerospike",
-    "functions": {
-        "bulkdelete": {
-            "function_name": "bulkdelete",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 5620
-        },
-        "bulkset": {
-            "function_name": "bulkset",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 6391
-        },
-        "delete": {
-            "function_name": "delete",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 719
-        },
-        "get": {
-            "function_name": "get",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 3603
-        },
-        "init": {
-            "function_name": "init",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 11485
-        },
-        "set": {
-            "function_name": "set",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 2242
-        }
-    }
-}
