--- tests/results/state/memcached.json
@@ -1,43 +1,0 @@
-{
-    "conformant": true,
-    "component_type": "state",
-    "component_name": "memcached",
-    "functions": {
-        "bulkdelete": {
-            "function_name": "bulkdelete",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 5458
-        },
-        "bulkset": {
-            "function_name": "bulkset",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 6710
-        },
-        "delete": {
-            "function_name": "delete",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 747
-        },
-        "get": {
-            "function_name": "get",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 579
-        },
-        "init": {
-            "function_name": "init",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 1388
-        },
-        "set": {
-            "function_name": "set",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 794
-        }
-    }
-}
