--- tests/results/state/mongodb.json
@@ -1,43 +1,0 @@
-{
-    "conformant": true,
-    "component_type": "state",
-    "component_name": "mongodb",
-    "functions": {
-        "bulkdelete": {
-            "function_name": "bulkdelete",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 20429
-        },
-        "bulkset": {
-            "function_name": "bulkset",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 21096
-        },
-        "delete": {
-            "function_name": "delete",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 2106
-        },
-        "get": {
-            "function_name": "get",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 666
-        },
-        "init": {
-            "function_name": "init",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 69
-        },
-        "set": {
-            "function_name": "set",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 6020
-        }
-    }
-}
