--- tests/results/state/redis.json
@@ -1,43 +1,0 @@
-{
-    "conformant": true,
-    "component_type": "state",
-    "component_name": "redis",
-    "functions": {
-        "bulkdelete": {
-            "function_name": "bulkdelete",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 6096
-        },
-        "bulkset": {
-            "function_name": "bulkset",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 6072
-        },
-        "delete": {
-            "function_name": "delete",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 715
-        },
-        "get": {
-            "function_name": "get",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 679
-        },
-        "init": {
-            "function_name": "init",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 2337
-        },
-        "set": {
-            "function_name": "set",
-            "included": true,
-            "conformed": true,
-            "execution_duration_in_μs": 781
-        }
-    }
-}
